requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
httpx==0.27.2
jinja2
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3

python-dotenv==1.0.1